pyproject.toml
[tool.poetry]
name = "pdappend"
version = "0.2.0"
description = "Append csv, xlsx, and xls files."
authors = ["Chris <cnpryer@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/cnpryer/pdappend"
documentation = "https://github.com/cnpryer/pdappend/wiki"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.1"
python-dotenv = "^0.19.2"
openpyxl = "^3.0.9"
xlrd = "^2.0.1"
click = "^8.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
black = "^22.1.0"
flake8 = "^4.0.1"
wheel = "^0.37.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pdappend = 'pdappend.cli:main'
pdappend-gui = 'pdappend.gui:main'